requirements.txt
flask==3.0.0
redis==5.0.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.5
numpy==1.26.0
pandas==2.1.1
faiss-cpu==1.7.4
requests==2.31.0
python-dotenv==1.0.0
openai==1.10.0
google-generativeai==0.3.2
